@@ -1,7 +1,7 @@
 [package]
 name = "dvc-data"
 version = "0.2.0"
-edition = "2021"
+edition = "2024"
 authors = ["Saugat Pachhai"]
 description = "Pure Rust implementation of DVC"
 repository = "https://github.com/skshetry/dvc-data-rs/"
@@ -15,37 +15,37 @@ eula = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-reflink-copy = "0.1.19"
+reflink-copy = "0.1.26"
 hex = "0.4.3"
-console = "0.15.8"
+console = "0.15.11"
 num_cpus = "1.16.0"
 md-5 = "0.10.6"
 rayon = "1.10.0"
 jwalk = "0.8.1"
 base16ct = { version = "0.2.0", features = ["alloc"] }
-clap = { version = "4.5.15", features = ["derive"] }
-indicatif = { version = "0.17.8", features = ["rayon"] }
-rusqlite = { version = "0.32.1", features = ["bundled", "chrono"] }
+clap = { version = "4.5.36", features = ["derive"] }
+indicatif = { version = "0.17.11", features = ["rayon"] }
+rusqlite = { version = "0.34.0", features = ["bundled", "chrono"] }
 serde = { version = "1.0", features = ["derive"] }
-serde_json = { version = "1.0.124", features = ["preserve_order"] }
+serde_json = { version = "1.0.140", features = ["preserve_order"] }
 serde_yaml = "0.9"
-ignore = "0.4.22"
-itertools = "0.13.0"
-env_logger = "0.11.5"
-log = "0.4.22"
-config = "0.14.0"
-serde_derive = "1.0.206"
-directories = "5.0.1"
-git2 = { version = "0.19.0", default-features = false }
-whoami = "1.5.1"
-serde_with = "3.9.0"
-thiserror = "1.0.63"
+ignore = "0.4.23"
+itertools = "0.14.0"
+env_logger = "0.11.8"
+log = "0.4.27"
+config = "0.15.11"
+serde_derive = "1.0.219"
+directories = "6.0.0"
+git2 = { version = "0.20.1", default-features = false }
+whoami = "1.6.0"
+serde_with = "3.12.0"
+thiserror = "2.0.12"
 
 [target.'cfg(windows)'.dependencies]
-file-id = "0.2.1"
+file-id = "0.2.2"
 
 [dev-dependencies]
-tempfile = "3.12.0"
+tempfile = "3.19.1"
 
 [lints.rust]
 unsafe_code = "forbid"
@@ -72,7 +72,13 @@ cargo-dist-version = "0.10.0"
 # The installers to generate for each app
 installers = ["shell", "powershell", "msi"]
 # Target platforms to build apps for (Rust target-triple syntax)
-targets = ["aarch64-apple-darwin", "x86_64-apple-darwin", "x86_64-unknown-linux-gnu", "x86_64-unknown-linux-musl", "x86_64-pc-windows-msvc"]
+targets = [
+    "aarch64-apple-darwin",
+    "x86_64-apple-darwin",
+    "x86_64-unknown-linux-gnu",
+    "x86_64-unknown-linux-musl",
+    "x86_64-pc-windows-msvc",
+]
 # CI backends to support
 ci = ["github"]
 # Publish jobs to run in CI
